lab2/consumer/requirements.txt
azure-eventhub
requests
asyncio
azure-storage-file-datalake